+++ lib/isl/src/main/kotlin/org/partiql/ionschema/discovery/TypeConstraint.kt
@@ -1,4 +1,4 @@
-package org.partiql.lang.schemadiscovery
+package org.partiql.ionschema.discovery
 
 import org.partiql.ionschema.model.IonSchemaModel