+++ partiql-lang/src/test/kotlin/org/partiql/lang/util/SchemaHelpers.kt
@@ -2,9 +2,9 @@ package org.partiql.lang.util
 
 import com.amazon.ion.IonSystem
 import com.amazon.ionschema.IonSchemaSystemBuilder
-import org.partiql.lang.partiqlisl.getResourceAuthority
+import org.partiql.lang.util.impl.ResourceAuthority
 
 fun createPartiqlIonSchemaSystem(ion: IonSystem) = IonSchemaSystemBuilder.standard()
-    .addAuthority(getResourceAuthority(ion))
+    .addAuthority(ResourceAuthority.getResourceAuthority(ion))
     .withIonSystem(ion)
     .build()